requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-cors-headers==4.7.0
django-oauth-toolkit==3.0.1
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.3.1
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
